Added Authorization to Endpoints

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
 import { CreateTaskDTO } from './dto/create-task.dto';
 import { GetTasksFilterDTO } from './dto/get-tasks-filter.dto';
 import { TaskStatus } from './tasks-status.enum';
@@ -14,12 +15,15 @@ export class TasksService {
         private taskRepository: TaskRepository
     ) {}
 
-    getTasks(filterDto: GetTasksFilterDTO): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    getTasks(
+        filterDto: GetTasksFilterDTO,
+        user: User
+    ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({where: { id, userId: user.id}});
 
         if (!found) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -27,20 +31,23 @@ export class TasksService {
         return found;
     }
 
-    async createTask(createTaskDTO: CreateTaskDTO) {
-        return this.taskRepository.createTask(createTaskDTO);
+    async createTask(
+        createTaskDTO: CreateTaskDTO,
+        user: User,
+    ): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDTO, user);
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = this.taskRepository.delete(id);
+    async deleteTask(id: number, user:User): Promise<void> {
+        const result = this.taskRepository.delete({id, userId: user.id});
         
         if((await result).affected == 0) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save()
         return task;
